+++ packages/backend/src/types.ts
@@ -9,16 +9,21 @@ import { WarehouseCatalog } from '@lightdash/warehouses';
 
 export interface ProjectAdapter {
     compileAllExplores(): Promise<(Explore | ExploreError)[]>;
+
     getDbtPackages(): Promise<DbtPackages | undefined>;
+
     test(): Promise<void>;
+
     destroy(): Promise<void>;
 }
 
 export interface DbtClient {
-    installDeps(): Promise<void>;
+    installDeps?(): Promise<void>;
+
     getDbtManifest(): Promise<DbtRpcGetManifestResults>;
-    getDbtCatalog(): Promise<DbtRpcDocsGenerateResults>;
+
     getDbtPackages?(): Promise<DbtPackages | undefined>;
+
     test(): Promise<void>;
 }